old-files/cobotpoc/container-images/mic/requirements.txt
numpy==1.26.4
scipy==1.11.4
sounddevice==0.4.6
soundfile==0.12.1
nemo_toolkit[asr]==2.7.2
